executor: move mv purge adaptive batch size test out of benchmark file
